Rename to providers.ts
