+++ mod.ts
@@ -1,5 +1,5 @@
 // Copyright 2023 the Deno authors. All rights reserved. MIT license.
-export * from "./src/create_oauth2_client.ts";
+export * from "./src/providers.ts";
 export * from "./src/get_session_access_token.ts";
 export * from "./src/handle_callback.ts";
 export * from "./src/get_session_id.ts";
